% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjanal.R
\name{conjanal}
\alias{conjanal}
\title{Conjoint analysis}
\usage{
conjanal(resp, profile, monotone = F, tor = 1e-10, maxit = 50)
}
\arguments{
\item{resp}{Response matrix where rows are profiles and columnes are responses from respondents.
Small value for less preferable and large value for more preferable.}

\item{profile}{Profile matrix where rows are profiles and columens are factors.}

\item{monotone}{TRUE for Kruskal monotonic transformation. FALSE for metric scale.}

\item{tor}{Tolerane for monotonic transformation.}

\item{maxit}{Maximum number of itneractions for monotonic transformation.}
}
\value{
A conjanal object
\item{lm }{lm object. Results of the linear regression models}
\item{part }{part worth for each level of each factor of each respondent}
\item{imp }{Relative importance of each factor for each respondent}
}
\description{
Metric and monotonic conjoint analysis for categorical factors
}
\details{
Metric conjoint analysis:\cr
Linear regression model \cr \cr
Monotonic conjoint analysis: \cr
Kruskal monotonic transformation is applied to each response column by opscale(level=2,...) in optiscale package.
Linear regression is applied to the transformed responses.
}
\examples{
library(conjanal)
contrasts(mydata$f1)
contrasts(mydata$f2)
contrasts(mydata$f3)

#define response matrix
resp<-as.matrix(mydata[,4:5])

#define profiles
profile<-mydata[,1:3]

fit.conj<-conjanal(resp=resp,profile=profile,monotone = TRUE)
summary(fit.conj$lm)

#part-worth
fit.conj$part

#relative importance
fit.conj$imp

#predicted utilities
cbind(profile,pred=predict(fit.conj$lm,newdata=profile))

#estimation of market shares
maxconj(fit.conj,csimp)
btlconj(fit.conj,csimp)
logitconj(fit.conj,csimp)
}
\author{
Chi-wai Kwan
}
