% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.conjanal.R
\name{logitconj}
\alias{logitconj}
\title{Logit method}
\usage{
logitconj(conjanal, simp)
}
\arguments{
\item{conjanal}{A conjanal object}

\item{simp}{A data frame of scenarios}
}
\value{
A matrix of scenarios and the estimated market shares
}
\description{
Logit method to estimate market shares
}
